+++ relay-metrics/src/aggregatorservice.rs
@@ -417,7 +417,7 @@ mod tests {
     use relay_common::time::UnixTimestamp;
     use relay_system::{FromMessage, Interface};
 
-    use crate::{BucketCountInquiry, BucketValue};
+    use crate::{BucketCountInquiry, BucketMetadata, BucketValue};
 
     use super::*;
 
@@ -478,6 +478,7 @@ mod tests {
             name: "c:transactions/foo".to_owned(),
             value: BucketValue::counter(42.into()),
             tags: BTreeMap::new(),
+            metadata: BucketMetadata::new(),
         }
     }