+++ relay-server/src/envelope.rs
@@ -597,7 +597,7 @@ impl AddAssign for SourceQuantities {
         } = self;
         *transactions += other.transactions;
         *profiles += other.profiles;
-        *buckets = other.buckets;
+        *buckets += other.buckets;
     }
 }