+++ relay-server/src/extractors/request_meta.rs
@@ -12,6 +12,7 @@ use axum::http::StatusCode;
 use axum::response::{IntoResponse, Response};
 use axum::RequestPartsExt;
 use data_encoding::BASE64;
+use relay_auth::RelayId;
 use relay_base_schema::project::{ParseProjectKeyError, ProjectId, ProjectKey};
 use relay_common::{Auth, Dsn, ParseAuthError, ParseDsnError, Scheme};
 use relay_config::UpstreamDescriptor;
@@ -224,10 +225,16 @@ pub struct RequestMeta<D = PartialDsn> {
     no_cache: bool,
 
     /// The time at which the request started.
-    //
-    // NOTE: This is internal-only and not exposed to Envelope headers.
+    ///
+    /// NOTE: This is internal-only and not exposed to Envelope headers.
     #[serde(skip, default = "Instant::now")]
     start_time: Instant,
+
+    /// Whether the request is coming from an statically configured internal Relay.
+    ///
+    /// NOTE: This is internal-only and not exposed to Envelope headers.
+    #[serde(skip)]
+    from_internal_relay: bool,
 }
 
 impl<D> RequestMeta<D> {
@@ -304,6 +311,11 @@ impl<D> RequestMeta<D> {
     pub fn set_start_time(&mut self, start_time: Instant) {
         self.start_time = start_time
     }
+
+    /// Whether the request is coming from a statically configured internal Relay.
+    pub fn is_from_internal_relay(&self) -> bool {
+        self.from_internal_relay
+    }
 }
 
 impl RequestMeta {
@@ -320,6 +332,7 @@ impl RequestMeta {
             no_cache: false,
             start_time: Instant::now(),
             client_hints: ClientHints::default(),
+            from_internal_relay: false,
         }
     }
 
@@ -421,6 +434,9 @@ impl PartialMeta {
         if self.user_agent.is_some() {
             complete.user_agent = self.user_agent;
         }
+        if self.from_internal_relay {
+            complete.from_internal_relay = self.from_internal_relay;
+        }
 
         complete.client_hints.copy_from(self.client_hints);
 
@@ -433,16 +449,32 @@ impl PartialMeta {
 }
 
 #[axum::async_trait]
-impl<S> FromRequestParts<S> for PartialMeta
-where
-    S: Send + Sync,
-{
+impl FromRequestParts<ServiceState> for PartialMeta {
     type Rejection = Infallible;
 
-    async fn from_request_parts(parts: &mut Parts, state: &S) -> Result<Self, Self::Rejection> {
+    async fn from_request_parts(
+        parts: &mut Parts,
+        state: &ServiceState,
+    ) -> Result<Self, Self::Rejection> {
         let mut ua = RawUserAgentInfo::default();
         for (key, value) in &parts.headers {
             ua.set_ua_field_from_header(key.as_str(), value.to_str().ok().map(str::to_string));
+        }
+
+        let mut from_internal_relay = false;
+        let relay_id = parts
+            .headers
+            .get("x-sentry-relay-id")
+            .and_then(|h| h.to_str().ok())
+            .and_then(|h| h.parse::<RelayId>().ok());
+
+        if let Some(relay_id) = relay_id {
+            relay_log::configure_scope(|s| s.set_tag("relay_id", relay_id));
+            from_internal_relay = state
+                .config()
+                .static_relays()
+                .get(&relay_id)
+                .map_or(false, |ri| ri.internal);
         }
 
         Ok(RequestMeta {
@@ -464,6 +496,7 @@ where
                 .await?
                 .into_inner(),
             client_hints: ua.client_hints,
+            from_internal_relay,
         })
     }
 }
@@ -589,7 +622,7 @@ impl FromRequestParts<ServiceState> for RequestMeta {
             parts.extract().await.map_err(BadEventMeta::BadProject)?;
 
         let auth = auth_from_parts(parts, store_path.sentry_key)?;
-        let partial_meta = parts.extract::<PartialMeta>().await?;
+        let partial_meta: PartialMeta = parts.extract_with_state(state).await?;
         let (public_key, key_flags) = ProjectKey::parse_with_flags(auth.public_key())?;
 
         let config = state.config();
@@ -626,6 +659,7 @@ impl FromRequestParts<ServiceState> for RequestMeta {
             no_cache: key_flags.contains(&"no-cache"),
             start_time: partial_meta.start_time,
             client_hints: partial_meta.client_hints,
+            from_internal_relay: partial_meta.from_internal_relay,
         })
     }
 }
@@ -648,6 +682,7 @@ mod tests {
                 no_cache: false,
                 start_time: Instant::now(),
                 client_hints: ClientHints::default(),
+                from_internal_relay: false,
             }
         }
     }
@@ -692,6 +727,7 @@ mod tests {
                 ),
                 sec_ch_ua_model: None,
             },
+            from_internal_relay: false,
         };
         deserialized.start_time = reqmeta.start_time;
         assert_eq!(deserialized, reqmeta);