+++ relay-server/src/metrics_extraction/generic.rs
@@ -1,8 +1,11 @@
 use std::collections::BTreeMap;
 
 use relay_common::time::UnixTimestamp;
-use relay_dynamic_config::{MetricExtractionConfig, Options, TagMapping, TagSource, TagSpec};
-use relay_metrics::{Bucket, BucketValue, FiniteF64, MetricResourceIdentifier, MetricType};
+use relay_dynamic_config::{MetricExtractionConfig, TagMapping, TagSource, TagSpec};
+
+use relay_metrics::{
+    Bucket, BucketMetadata, BucketValue, FiniteF64, MetricResourceIdentifier, MetricType,
+};
 use relay_protocol::{Getter, Val};
 use relay_quotas::DataCategory;
 
@@ -20,11 +23,7 @@ pub trait Extractable: Getter {
 /// The instance must have a valid timestamp; if the timestamp is missing or invalid, no metrics are
 /// extracted. Timestamp and clock drift correction should occur before metrics extraction to ensure
 /// valid timestamps.
-pub fn extract_metrics<T>(
-    instance: &T,
-    config: &MetricExtractionConfig,
-    global_options: Option<&Options>,
-) -> Vec<Bucket>
+pub fn extract_metrics<T>(instance: &T, config: &MetricExtractionConfig) -> Vec<Bucket>
 where
     T: Extractable,
 {
@@ -35,16 +34,7 @@ where
         return metrics;
     };
 
-    // HACK: The killswitch for the usage metric has a different life cycle
-    // than the project config, so we cannot apply it in `ProjectConfig::sanitize`,
-    // which runs when the project config is updated.
-    // This hack can be removed once the usage metric is stable.
-    let allow_span_usage_metric = global_options.map_or(false, |options| options.span_usage_metric);
     for metric_spec in &config.metrics {
-        if !allow_span_usage_metric && metric_spec.mri == "c:spans/usage@none" {
-            continue;
-        }
-
         if metric_spec.category != instance.category() {
             continue;
         }
@@ -72,6 +62,7 @@ where
             value,
             timestamp,
             tags: extract_tags(instance, &metric_spec.tags),
+            metadata: BucketMetadata::new(),
         });
     }
 
@@ -192,7 +183,7 @@ mod tests {
         });
         let config = serde_json::from_value(config_json).unwrap();
 
-        let metrics = extract_metrics(event.value().unwrap(), &config, None);
+        let metrics = extract_metrics(event.value().unwrap(), &config);
         insta::assert_debug_snapshot!(metrics, @r###"
         [
             Bucket {
@@ -203,6 +194,9 @@ mod tests {
                     1.0,
                 ),
                 tags: {},
+                metadata: BucketMetadata {
+                    merges: 1,
+                },
             },
         ]
         "###);
@@ -229,7 +223,7 @@ mod tests {
         });
         let config = serde_json::from_value(config_json).unwrap();
 
-        let metrics = extract_metrics(event.value().unwrap(), &config, None);
+        let metrics = extract_metrics(event.value().unwrap(), &config);
         insta::assert_debug_snapshot!(metrics, @r###"
         [
             Bucket {
@@ -242,6 +236,9 @@ mod tests {
                     ],
                 ),
                 tags: {},
+                metadata: BucketMetadata {
+                    merges: 1,
+                },
             },
         ]
         "###);
@@ -270,7 +267,7 @@ mod tests {
         });
         let config = serde_json::from_value(config_json).unwrap();
 
-        let metrics = extract_metrics(event.value().unwrap(), &config, None);
+        let metrics = extract_metrics(event.value().unwrap(), &config);
         insta::assert_debug_snapshot!(metrics, @r###"
         [
             Bucket {
@@ -283,6 +280,9 @@ mod tests {
                     },
                 ),
                 tags: {},
+                metadata: BucketMetadata {
+                    merges: 1,
+                },
             },
         ]
         "###);
@@ -323,7 +323,7 @@ mod tests {
         });
         let config = serde_json::from_value(config_json).unwrap();
 
-        let metrics = extract_metrics(event.value().unwrap(), &config, None);
+        let metrics = extract_metrics(event.value().unwrap(), &config);
         insta::assert_debug_snapshot!(metrics, @r###"
         [
             Bucket {
@@ -337,6 +337,9 @@ mod tests {
                     "fast": "no",
                     "id": "4711",
                     "release": "myapp@1.0.0",
+                },
+                metadata: BucketMetadata {
+                    merges: 1,
                 },
             },
         ]
@@ -377,7 +380,7 @@ mod tests {
         });
         let config = serde_json::from_value(config_json).unwrap();
 
-        let metrics = extract_metrics(event.value().unwrap(), &config, None);
+        let metrics = extract_metrics(event.value().unwrap(), &config);
         insta::assert_debug_snapshot!(metrics, @r###"
         [
             Bucket {
@@ -389,6 +392,9 @@ mod tests {
                 ),
                 tags: {
                     "fast": "yes",
+                },
+                metadata: BucketMetadata {
+                    merges: 1,
                 },
             },
         ]
@@ -433,7 +439,7 @@ mod tests {
         });
         let config = serde_json::from_value(config_json).unwrap();
 
-        let metrics = extract_metrics(event.value().unwrap(), &config, None);
+        let metrics = extract_metrics(event.value().unwrap(), &config);
         insta::assert_debug_snapshot!(metrics, @r###"
         [
             Bucket {
@@ -445,6 +451,9 @@ mod tests {
                 ),
                 tags: {
                     "fast": "yes",
+                },
+                metadata: BucketMetadata {
+                    merges: 1,
                 },
             },
         ]
@@ -497,7 +506,7 @@ mod tests {
         });
         let config = serde_json::from_value(config_json).unwrap();
 
-        let metrics = extract_metrics(event.value().unwrap(), &config, None);
+        let metrics = extract_metrics(event.value().unwrap(), &config);
         insta::assert_debug_snapshot!(metrics, @r###"
         [
             Bucket {
@@ -510,6 +519,9 @@ mod tests {
                     ],
                 ),
                 tags: {},
+                metadata: BucketMetadata {
+                    merges: 1,
+                },
             },
         ]
         "###);